requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
alpha_vantage==3.0.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
certifi==2024.8.30
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
faiss-cpu==1.9.0.post1
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
googleapis-common-protos==1.66.0
groq==0.12.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-community==0.3.7
langchain-core==0.3.19
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
Lazify==0.4.0
literalai==0.0.623
lxml==4.9.4
MarkupSafe==3.0.2
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
ollama==0.3.3
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.11
packaging==23.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.10.0
PyMuPDF==1.24.14
pytest==8.3.3
pytest-asyncio==0.23.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.2
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
sympy==1.13.1
syncer==2.0.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.28.2
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.17.2
zipp==3.21.0
